requirements.txt
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
awsebcli==3.20.3
bcrypt==4.0.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
blessed==1.19.1
boto3==1.26.64
botocore==1.29.64
cached-property==1.5.2
celery==5.2.7
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.31.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
DateTime==4.4
decouple==0.0.7
Deprecated==1.2.13
distlib==0.3.5
dj-database-url==0.5.0
Django==3.2.13
django-appconf==1.0.5
django-baton==2.4.1
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-celery==3.1.17
django-celery-beat==2.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-custom-user-migration==0.3.0
django-debug-toolbar==3.6.0
django-docs==0.3.2
django-dotenv==1.4.2
django-environ==0.9.0
django-environ-2==2.3.0
django-extensions==3.1.5
django-flower==1.0.0
django-guid==3.3.0
django-heroku==0.3.1
django-imagekit==4.1.0
django-jazzmin==2.6.0
django-jinja==2.10.2
django-on-heroku==1.1.2
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-seed==0.3.1
django-storages==1.13.2
django-timezone-field==5.0
django-twilio==0.13.2
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.4
environ==1.0
exponent-server-sdk==2.0.0
Faker==15.3.4
filelock==3.8.0
flower==1.2.0
future==0.16.0
gevent==21.12.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.3
gunicorn==20.1.0
humanize==0.5.1
idna==3.3
imagekitio==3.0.1
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jotform @ git+https://github.com/jotform/jotform-api-python.git@050e92db72c315496e335be4959307e6a7043ed2
jsonschema==3.2.0
kombu==5.2.4
MarkupSafe==2.1.1
model-bakery==1.9.0
mysqlclient==2.1.1
numpy==1.23.3
packaging==21.3
paramiko==2.12.0
pathspec==0.9.0
phonenumbers==8.12.50
phonenumberslite==8.13.3
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
plumbum==1.7.2
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
psutil==5.9.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydotplus==2.0.2
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
redis==4.3.4
redislite==6.2.805324
requests==2.28.0
requests-toolbelt==0.9.1
rollbar==0.13.14
rpyc==5.2.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Rx==1.6.1
s3transfer==0.6.0
semantic-version==2.8.5
sentry-sdk==1.9.8
singledispatch==3.7.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.2
sshtunnel==0.4.0
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
toposort==1.7
tornado==6.2
twilio==6.63.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.3
wcwidth==0.2.5
websocket-client==0.59.0
whitenoise==6.2.0
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
